+++ .gitignore
@@ -1,12 +1,12 @@
 # dependencies
-*/node_modules
-*/.pnp
-*/.pnp.*
-*/.yarn/*
-*/!.yarn/patches
-*/!.yarn/plugins
-*/!.yarn/releases
-*/!.yarn/versions
+node_modules
+.pnp
+.pnp.*
+.yarn/*
+!.yarn/patches
+!.yarn/plugins
+!.yarn/releases
+!.yarn/versions
 
 # production
 */build
