+++ apps/next-app/next.config.ts
@@ -1,0 +1,7 @@
+import type { NextConfig } from 'next';
+
+const nextConfig: NextConfig = {
+	/* config options here */
+};
+
+export default nextConfig;
