+++ apps/next-app/package.json
@@ -1,0 +1,30 @@
+{
+	"name": "next-app",
+	"version": "0.1.0",
+	"private": true,
+	"scripts": {
+		"dev": "next dev --turbopack",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint",
+		"openapi-gen": "openapi-typescript ../../../schema/openapi.yaml -o ./generated/schema.d.ts"
+	},
+	"dependencies": {
+		"next": "15.2.2",
+		"openapi-fetch": "^0.13.5",
+		"react": "^19.0.0",
+		"react-dom": "^19.0.0"
+	},
+	"devDependencies": {
+		"@eslint/eslintrc": "^3",
+		"@tailwindcss/postcss": "^4",
+		"@types/node": "^20",
+		"@types/react": "^19",
+		"@types/react-dom": "^19",
+		"eslint": "^9",
+		"eslint-config-next": "15.2.2",
+		"openapi-typescript": "^7.6.1",
+		"tailwindcss": "^4",
+		"typescript": "catalog:"
+	}
+}
