+++ apps/next-app/src/app/favicon.ico
