--- js/apps/next-app/src/app/page.tsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <h1>Hello World!</h1>;
-}
