+++ package.json
@@ -1,0 +1,17 @@
+{
+	"name": "canalia",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"dev": "pnpm -C apps/next-app dev",
+		"openapi-gen": "pnpm -C apps/next-app openapi-gen"
+	},
+	"keywords": [],
+	"author": "s-union",
+	"license": "ISC",
+	"packageManager": "pnpm@10.5.2",
+	"devDependencies": {
+		"@biomejs/biome": "^1.9.4"
+	}
+}
